Upgrade Terraform to 0.10
Given coming Terraform 0.10 as well as the fact we want to start using
split provider plugins right away (as we've some contributed code that
we want to start using now, without having to wait for the next provider
plugin release), I upgraded the test-suite for tf_frontend_router
Terraform module to work with Terraform 0.10.

Quite a few changes, as 0.10 outputs plan results in a bit different
form than before.